+++ CHANGES.md
@@ -1,8 +1,8 @@
 # Changelog of clean-python
 
 
-## 0.4.2 (unreleased)
----------------------
+0.4.2 (2023-09-04)
+------------------
 
 - Support adding/changing `responses` via route_options.